Move to gulp 4
And while we're here, tidy up some dependencies.

Note that gulp-gh-page's maintainer has updated the things we were
relying on the `tekd` repo for, so we're switching back to that.
However, note that they haven't redeployed it to npm yet, so we're
setting the version in our package.json for now.