Github Actions etc
Move the page to build on Github Actions. This should, hopefully,
also resolve an issue in which pulls from Contentful were not
happening.

This commit currently pushes to test-branch and not to master
as a way of making sure that the build job works. A later commit
will finalize the deploy.